[data.search] Move background session service to data enhanced plugin (#84837)
* [data.search] Move search method inside session service and add tests

* Move background session service to data_enhanced plugin

* Fix types
