+++ docs/development/plugins/data/server/kibana-plugin-plugins-data-server.isessionservice.asscopedprovider.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [kibana-plugin-plugins-data-server](./kibana-plugin-plugins-data-server.md) &gt; [ISessionService](./kibana-plugin-plugins-data-server.isessionservice.md) &gt; [asScopedProvider](./kibana-plugin-plugins-data-server.isessionservice.asscopedprovider.md)
+
+## ISessionService.asScopedProvider property
+
+<b>Signature:</b>
+
+```typescript
+asScopedProvider: (core: CoreStart) => (request: KibanaRequest) => IScopedSessionService;
+```
