+++ docs/development/plugins/data/server/kibana-plugin-plugins-data-server.sessionservice._constructor_.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [kibana-plugin-plugins-data-server](./kibana-plugin-plugins-data-server.md) &gt; [SessionService](./kibana-plugin-plugins-data-server.sessionservice.md) &gt; [(constructor)](./kibana-plugin-plugins-data-server.sessionservice._constructor_.md)
+
+## SessionService.(constructor)
+
+Constructs a new instance of the `SessionService` class
+
+<b>Signature:</b>
+
+```typescript
+constructor();
+```
