+++ docs/development/plugins/data/server/kibana-plugin-plugins-data-server.sessionservice.asscopedprovider.md
@@ -1,0 +1,26 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [kibana-plugin-plugins-data-server](./kibana-plugin-plugins-data-server.md) &gt; [SessionService](./kibana-plugin-plugins-data-server.sessionservice.md) &gt; [asScopedProvider](./kibana-plugin-plugins-data-server.sessionservice.asscopedprovider.md)
+
+## SessionService.asScopedProvider() method
+
+<b>Signature:</b>
+
+```typescript
+asScopedProvider(core: CoreStart): (request: KibanaRequest) => {
+        search: <Request_1 extends IKibanaSearchRequest<any>, Response_1 extends IKibanaSearchResponse<any>>(strategy: ISearchStrategy<Request_1, Response_1>, request: Request_1, options: import("../../../common").ISearchOptions, deps: import("../types").SearchStrategyDependencies) => import("rxjs").Observable<Response_1>;
+    };
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  core | <code>CoreStart</code> |  |
+
+<b>Returns:</b>
+
+`(request: KibanaRequest) => {
+        search: <Request_1 extends IKibanaSearchRequest<any>, Response_1 extends IKibanaSearchResponse<any>>(strategy: ISearchStrategy<Request_1, Response_1>, request: Request_1, options: import("../../../common").ISearchOptions, deps: import("../types").SearchStrategyDependencies) => import("rxjs").Observable<Response_1>;
+    }`
+
