+++ docs/development/plugins/data/server/kibana-plugin-plugins-data-server.sessionservice.md
@@ -1,0 +1,27 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [kibana-plugin-plugins-data-server](./kibana-plugin-plugins-data-server.md) &gt; [SessionService](./kibana-plugin-plugins-data-server.sessionservice.md)
+
+## SessionService class
+
+The OSS session service. See data\_enhanced in X-Pack for the background session service.
+
+<b>Signature:</b>
+
+```typescript
+export declare class SessionService implements ISessionService 
+```
+
+## Constructors
+
+|  Constructor | Modifiers | Description |
+|  --- | --- | --- |
+|  [(constructor)()](./kibana-plugin-plugins-data-server.sessionservice._constructor_.md) |  | Constructs a new instance of the <code>SessionService</code> class |
+
+## Methods
+
+|  Method | Modifiers | Description |
+|  --- | --- | --- |
+|  [asScopedProvider(core)](./kibana-plugin-plugins-data-server.sessionservice.asscopedprovider.md) |  |  |
+|  [search(strategy, args)](./kibana-plugin-plugins-data-server.sessionservice.search.md) |  |  |
+
