+++ docs/development/plugins/data/server/kibana-plugin-plugins-data-server.sessionservice.search.md
@@ -1,0 +1,23 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [kibana-plugin-plugins-data-server](./kibana-plugin-plugins-data-server.md) &gt; [SessionService](./kibana-plugin-plugins-data-server.sessionservice.md) &gt; [search](./kibana-plugin-plugins-data-server.sessionservice.search.md)
+
+## SessionService.search() method
+
+<b>Signature:</b>
+
+```typescript
+search<Request extends IKibanaSearchRequest, Response extends IKibanaSearchResponse>(strategy: ISearchStrategy<Request, Response>, ...args: Parameters<ISearchStrategy<Request, Response>['search']>): import("rxjs").Observable<Response>;
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  strategy | <code>ISearchStrategy&lt;Request, Response&gt;</code> |  |
+|  args | <code>Parameters&lt;ISearchStrategy&lt;Request, Response&gt;['search']&gt;</code> |  |
+
+<b>Returns:</b>
+
+`import("rxjs").Observable<Response>`
+
