+++ src/plugins/data/common/search/index.ts
@@ -23,5 +23,4 @@ export * from './expressions';
 export * from './search_source';
 export * from './tabify';
 export * from './types';
-export * from './session';
 export * from './utils';
