+++ src/plugins/data/server/saved_objects/index.ts
@@ -20,4 +20,3 @@ export { querySavedObjectType } from './query';
 export { indexPatternSavedObjectType } from './index_patterns';
 export { kqlTelemetry } from './kql_telemetry';
 export { searchTelemetry } from './search_telemetry';
-export { BACKGROUND_SESSION_TYPE, backgroundSessionMapping } from './background_session';
