+++ src/plugins/data/server/search/session/types.ts
@@ -17,5 +17,19 @@
  * under the License.
  */
 
-export * from './status';
-export * from './types';
+import { Observable } from 'rxjs';
+import { CoreStart, KibanaRequest } from 'kibana/server';
+import { ISearchStrategy } from '../types';
+import { IKibanaSearchRequest, IKibanaSearchResponse } from '../../../common/search';
+
+export interface IScopedSessionService {
+  search: <Request extends IKibanaSearchRequest, Response extends IKibanaSearchResponse>(
+    strategy: ISearchStrategy<Request, Response>,
+    ...args: Parameters<ISearchStrategy<Request, Response>['search']>
+  ) => Observable<Response>;
+  [prop: string]: any;
+}
+
+export interface ISessionService {
+  asScopedProvider: (core: CoreStart) => (request: KibanaRequest) => IScopedSessionService;
+}
