+++ x-pack/plugins/data_enhanced/common/search/session/index.ts
@@ -4,6 +4,5 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+export * from './status';
 export * from './types';
-export * from './poll_search';
-export * from './session';
