+++ x-pack/plugins/data_enhanced/common/search/session/status.ts
@@ -4,6 +4,10 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export * from './types';
-export * from './poll_search';
-export * from './session';
+export enum BackgroundSessionStatus {
+  IN_PROGRESS = 'in_progress',
+  ERROR = 'error',
+  COMPLETE = 'complete',
+  CANCELLED = 'cancelled',
+  EXPIRED = 'expired',
+}
