+++ x-pack/plugins/data_enhanced/server/routes/session.test.ts
@@ -1,27 +1,14 @@
 /*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
  */
 
 import type { MockedKeys } from '@kbn/utility-types/jest';
 import type { CoreSetup, RequestHandlerContext } from 'kibana/server';
-import type { DataPluginStart } from '../../plugin';
-import { coreMock, httpServerMock } from '../../../../../../src/core/server/mocks';
-import { createSearchRequestHandlerContext } from '../mocks';
+import { coreMock, httpServerMock } from '../../../../../src/core/server/mocks';
+import { PluginStart as DataPluginStart } from '../../../../../src/plugins/data/server';
+import { createSearchRequestHandlerContext } from './mocks';
 import { registerSessionRoutes } from './session';
 
 describe('registerSessionRoutes', () => {
