+++ x-pack/plugins/data_enhanced/server/search/index.ts
@@ -6,3 +6,4 @@
 
 export { enhancedEsSearchStrategyProvider } from './es_search_strategy';
 export { eqlSearchStrategyProvider } from './eql_search_strategy';
+export * from './session';
